requirements.txt
absl-py==0.7.1
aniso8601==7.0.0
asn1crypto==0.24.0
astor==0.8.0
astroid==2.2.5
attrs==19.1.0
backcall==0.1.0
bleach==3.1.0
botocore==1.12.130
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
checkpoints==0.0.1
Click==7.0
colorama==0.3.9
cryptography==2.7
cycler==0.10.0
cython==0.29.13
decorator==4.4.0
defusedxml==0.6.0
docutils==0.14
entrypoints==0.3
Flask==1.0.2
Flask-RESTful==0.3.7
Flask-SocketIO==3.3.2
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
gast==0.2.2
google-pasta==0.1.7
grpcio==1.22.0
gunicorn==19.9.0
h5py==2.9.0
idna==2.8
inflection==0.3.1
ipykernel==5.1.1
ipython==7.6.1
ipython-genutils==0.2.0
isort==4.3.20
itsdangerous==1.1.0
jedi==0.14.1
Jinja2==2.10
jmespath==0.9.4
joblib==0.13.2
json5==0.8.5
jsonschema==3.0.1
jupyter-client==5.3.1
jupyter-core==4.5.0
jupyterlab==1.0.2
jupyterlab-server==1.0.0
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
lazy-object-proxy==1.4.1
lxml==4.4.0
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.1
mccabe==0.6.1
mistune==0.8.4
more-itertools==5.0.0
nbconvert==5.5.0
nbformat==4.4.0
ndg-httpsclient==0.5.1
ngrok==0.0.1
nltk==3.4.4
notebook==5.7.8
numpy==1.16.4
oauthlib==3.0.2
opencv-python==4.1.0.25
pandas==0.24.2
pandocfilters==1.4.2
parso==0.5.1
pickleshare==0.7.5
Pillow==6.1.0
pipenv==2018.11.26
prometheus-client==0.7.1
prompt-toolkit==2.0.9
protobuf==3.9.0
pyasn1==0.4.5
pycparser==2.19
Pygments==2.4.2
pylint==2.3.1
pyOpenSSL==19.0.0
pyparsing==2.4.0
pyrsistent==0.15.3
PySocks==1.7.0
python-dateutil==2.8.0
python-dotenv==0.10.3
python-engineio==3.5.1
python-socketio==3.1.2
pytz==2018.9
PyYAML==3.13
pyzmq==18.0.2
Quandl==3.4.8
ratelim==0.1.6
requests==2.21.0
requests-oauthlib==1.2.0
rsa==3.4.2
s3transfer==0.2.0
scikit-learn==0.21.2
scipy==1.3.0
Send2Trash==1.5.0
six==1.12.0
SQLAlchemy==1.3.2
tb-nightly==1.14.0a20190603
tensorboard==1.14.0
tensorflow==2.0.0b1
tensorflow-estimator==1.14.0
termcolor==1.1.0
terminado==0.8.2
testpath==0.4.2
tf-estimator-nightly==1.14.0.dev2019060501
tornado==6.0.3
tqdm==4.32.2
traitlets==4.3.2
tweepy==3.8.0
typed-ast==1.4.0
urllib3==1.24.1
virtualenv==16.6.1
virtualenv-clone==0.5.3
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.15.2
wrapt==1.11.2
WTForms==2.2.1
